+++ src/worker_pool.cpp
@@ -2,7 +2,7 @@
 #include <functional>
 
 void WorkerPool::start() {
-  for (uint32_t i = 0; i < num_threads_; i++) {
+  for (size_t i = 0; i < num_threads_; i++) {
     std::shared_ptr<JobWorker> worker(new JobWorker(job_queue_, i + 1));
     std::shared_ptr<std::thread> t(
         new std::thread(std::bind(&JobWorker::run, worker)));
@@ -24,7 +24,7 @@ bool WorkerPool::is_empty() {
   return job_queue_->is_empty();
 }
 
-uint32_t WorkerPool::count_jobs() {
+size_t WorkerPool::count_jobs() {
   return job_queue_->count();
 }
 
@@ -32,7 +32,7 @@ void WorkerPool::stop_plz() {
   // clear all queued jobs
   job_queue_->clear();
   // enqueue NullJob to stop threads
-  for (uint32_t i = 0; i < num_threads_; i++) {
+  for (size_t i = 0; i < num_threads_; i++) {
     JobItemPtr job(new NullJob());
     queue_job(job);
   }
